v4l2_wrapper: Split into 2 classes
Refactor v4l2_wrapper class, split into 2 classes:

* V4L2Wrapper - is implementing abstract interface for rest of the
HAL. Class also provides default implementation for V4L2 controls setting/getting,
formats checking, etc. Any buffer management should be implemented by child classes.
NewV4L2Wrapper should choose proper child implementation depending on the specified
device path.

* V4L2GenericWrapper - implements original buffers management based on V4L2 userptr
buffers.
