+++ common/camera/v4l2_wrapper.cpp
@@ -18,6 +18,7 @@
 #define LOG_TAG "V4L2Wrapper"
 
 #include "v4l2_wrapper.h"
+#include "v4l2_generic_wrapper.h"
 
 #include <algorithm>
 #include <array>
@@ -36,7 +37,6 @@
 
 namespace v4l2_camera_hal {
 
-using arc::AllocatedFrameBuffer;
 using arc::SupportedFormat;
 using arc::SupportedFormats;
 using default_camera_hal::CaptureRequest;
@@ -56,7 +56,7 @@ const int32_t kStandardSizes[][2] = {
 };
 
 V4L2Wrapper* V4L2Wrapper::NewV4L2Wrapper(const std::string device_path) {
-  return new V4L2Wrapper(device_path);
+  return new V4L2GenericWrapper(device_path);
 }
 
 V4L2Wrapper::V4L2Wrapper(const std::string device_path)
@@ -121,23 +121,6 @@ void V4L2Wrapper::Disconnect() {
 
   device_fd_.reset(-1);  // Includes close().
   format_.reset();
-  {
-    std::lock_guard<std::mutex> buffer_lock(buffer_queue_lock_);
-    buffers_.clear();
-  }
-}
-
-// Helper function. Should be used instead of ioctl throughout this class.
-template <typename T>
-int V4L2Wrapper::IoctlLocked(int request, T data) {
-  // Potentially called so many times logging entry is a bad idea.
-  std::lock_guard<std::mutex> lock(device_lock_);
-
-  if (!connected()) {
-    HAL_LOGE("Device %s not connected.", device_path_.c_str());
-    return -ENODEV;
-  }
-  return TEMP_FAILURE_RETRY(ioctl(device_fd_.get(), request, data));
 }
 
 int V4L2Wrapper::StreamOn() {
@@ -170,11 +153,6 @@ int V4L2Wrapper::StreamOff() {
   if (res < 0) {
     HAL_LOGE("STREAMOFF fails: %s", strerror(errno));
     return -ENODEV;
-  }
-  std::lock_guard<std::mutex> lock(buffer_queue_lock_);
-  for (auto& buffer : buffers_) {
-    buffer.active = false;
-    buffer.request.reset();
   }
   HAL_LOGV("Stream turned off.");
   return 0;
@@ -491,251 +469,6 @@ int V4L2Wrapper::GetFormatFrameDurationRange(
   (*duration_range)[0] = min;
   (*duration_range)[1] = max;
   return 0;
-}
-
-int V4L2Wrapper::SetFormat(const StreamFormat& desired_format,
-                           uint32_t* result_max_buffers) {
-  HAL_LOG_ENTER();
-
-  if (format_ && desired_format == *format_) {
-    HAL_LOGV("Already in correct format, skipping format setting.");
-    *result_max_buffers = buffers_.size();
-    return 0;
-  }
-
-  if (format_) {
-    // If we had an old format, first request 0 buffers to inform the device
-    // we're no longer using any previously "allocated" buffers from the old
-    // format. This seems like it shouldn't be necessary for USERPTR memory,
-    // and/or should happen from turning the stream off, but the driver
-    // complained. May be a driver issue, or may be intended behavior.
-    int res = RequestBuffers(0);
-    if (res) {
-      return res;
-    }
-  }
-
-  // Select the matching format, or if not available, select a qualified format
-  // we can convert from.
-  SupportedFormat format;
-  if (!StreamFormat::FindBestFitFormat(supported_formats_, qualified_formats_,
-                                       desired_format.v4l2_pixel_format(),
-                                       desired_format.width(),
-                                       desired_format.height(), &format)) {
-    HAL_LOGE(
-        "Unable to find supported resolution in list, "
-        "width: %d, height: %d",
-        desired_format.width(), desired_format.height());
-    return -EINVAL;
-  }
-
-  // Set the camera to the new format.
-  v4l2_format new_format;
-  const StreamFormat resolved_format(format);
-  resolved_format.FillFormatRequest(&new_format);
-
-  // TODO(b/29334616): When async, this will need to check if the stream
-  // is on, and if so, lock it off while setting format.
-  if (IoctlLocked(VIDIOC_S_FMT, &new_format) < 0) {
-    HAL_LOGE("S_FMT failed: %s", strerror(errno));
-    return -ENODEV;
-  }
-
-  // Check that the driver actually set to the requested values.
-  if (resolved_format != new_format) {
-    HAL_LOGE("Device doesn't support desired stream configuration.");
-    return -EINVAL;
-  }
-
-  // Keep track of our new format.
-  format_.reset(new StreamFormat(new_format));
-
-  // Format changed, request new buffers.
-  int res = RequestBuffers(1);
-  if (res) {
-    HAL_LOGE("Requesting buffers for new format failed.");
-    return res;
-  }
-  *result_max_buffers = buffers_.size();
-  return 0;
-}
-
-int V4L2Wrapper::RequestBuffers(uint32_t num_requested) {
-  v4l2_requestbuffers req_buffers;
-  memset(&req_buffers, 0, sizeof(req_buffers));
-  req_buffers.type = format_->type();
-  req_buffers.memory = V4L2_MEMORY_USERPTR;
-  req_buffers.count = num_requested;
-
-  int res = IoctlLocked(VIDIOC_REQBUFS, &req_buffers);
-  // Calling REQBUFS releases all queued buffers back to the user.
-  if (res < 0) {
-    HAL_LOGE("REQBUFS failed: %s", strerror(errno));
-    return -ENODEV;
-  }
-
-  // V4L2 will set req_buffers.count to a number of buffers it can handle.
-  if (num_requested > 0 && req_buffers.count < 1) {
-    HAL_LOGE("REQBUFS claims it can't handle any buffers.");
-    return -ENODEV;
-  }
-  buffers_.resize(req_buffers.count);
-  return 0;
-}
-
-int V4L2Wrapper::EnqueueRequest(
-    std::shared_ptr<default_camera_hal::CaptureRequest> request) {
-  if (!format_) {
-    HAL_LOGE("Stream format must be set before enqueuing buffers.");
-    return -ENODEV;
-  }
-
-  // Find a free buffer index. Could use some sort of persistent hinting
-  // here to improve expected efficiency, but buffers_.size() is expected
-  // to be low enough (<10 experimentally) that it's not worth it.
-  int index = -1;
-  {
-    std::lock_guard<std::mutex> guard(buffer_queue_lock_);
-    for (size_t i = 0; i < buffers_.size(); ++i) {
-      if (!buffers_[i].active) {
-        index = i;
-        break;
-      }
-    }
-  }
-  if (index < 0) {
-    // Note: The HAL should be tracking the number of buffers in flight
-    // for each stream, and should never overflow the device.
-    HAL_LOGE("Cannot enqueue buffer: stream is already full.");
-    return -ENODEV;
-  }
-
-  // Set up a v4l2 buffer struct.
-  v4l2_buffer device_buffer;
-  memset(&device_buffer, 0, sizeof(device_buffer));
-  device_buffer.type = format_->type();
-  device_buffer.index = index;
-
-  // Use QUERYBUF to ensure our buffer/device is in good shape,
-  // and fill out remaining fields.
-  if (IoctlLocked(VIDIOC_QUERYBUF, &device_buffer) < 0) {
-    HAL_LOGE("QUERYBUF fails: %s", strerror(errno));
-    // Return buffer index.
-    std::lock_guard<std::mutex> guard(buffer_queue_lock_);
-    buffers_[index].active = false;
-    return -ENODEV;
-  }
-
-  // Setup our request context and fill in the user pointer field.
-  RequestContext* request_context;
-  void* data;
-  {
-    std::lock_guard<std::mutex> guard(buffer_queue_lock_);
-    request_context = &buffers_[index];
-    request_context->camera_buffer->SetDataSize(device_buffer.length);
-    request_context->camera_buffer->Reset();
-    request_context->camera_buffer->SetFourcc(format_->v4l2_pixel_format());
-    request_context->camera_buffer->SetWidth(format_->width());
-    request_context->camera_buffer->SetHeight(format_->height());
-    request_context->request = request;
-    data = request_context->camera_buffer->GetData();
-  }
-  device_buffer.m.userptr = reinterpret_cast<unsigned long>(data);
-
-  // Pass the buffer to the camera.
-  if (IoctlLocked(VIDIOC_QBUF, &device_buffer) < 0) {
-    HAL_LOGE("QBUF fails: %s", strerror(errno));
-    return -ENODEV;
-  }
-
-  // Mark the buffer as in flight.
-  std::lock_guard<std::mutex> guard(buffer_queue_lock_);
-  request_context->active = true;
-
-  return 0;
-}
-
-int V4L2Wrapper::DequeueRequest(std::shared_ptr<CaptureRequest>* request) {
-  if (!format_) {
-    HAL_LOGV(
-        "Format not set, so stream can't be on, "
-        "so no buffers available for dequeueing");
-    return -EAGAIN;
-  }
-
-  v4l2_buffer buffer;
-  memset(&buffer, 0, sizeof(buffer));
-  buffer.type = format_->type();
-  buffer.memory = V4L2_MEMORY_USERPTR;
-  int res = IoctlLocked(VIDIOC_DQBUF, &buffer);
-  if (res) {
-    if (errno == EAGAIN) {
-      // Expected failure.
-      return -EAGAIN;
-    } else {
-      // Unexpected failure.
-      HAL_LOGE("DQBUF fails: %s", strerror(errno));
-      return -ENODEV;
-    }
-  }
-
-  std::lock_guard<std::mutex> guard(buffer_queue_lock_);
-  RequestContext* request_context = &buffers_[buffer.index];
-
-  // Lock the camera stream buffer for painting.
-  const camera3_stream_buffer_t* stream_buffer =
-      &request_context->request->output_buffers[0];
-  uint32_t fourcc =
-      StreamFormat::HalToV4L2PixelFormat(stream_buffer->stream->format);
-
-  if (request) {
-    *request = request_context->request;
-  }
-
-  // Note that the device buffer length is passed to the output frame. If the
-  // GrallocFrameBuffer does not have support for the transformation to
-  // |fourcc|, it will assume that the amount of data to lock is based on
-  // |buffer.length|, otherwise it will use the ImageProcessor::ConvertedSize.
-  arc::GrallocFrameBuffer output_frame(
-      *stream_buffer->buffer, stream_buffer->stream->width,
-      stream_buffer->stream->height, fourcc, buffer.length,
-      stream_buffer->stream->usage);
-  res = output_frame.Map();
-  if (res) {
-    HAL_LOGE("Failed to map output frame.");
-    request_context->request.reset();
-    return -EINVAL;
-  }
-  if (request_context->camera_buffer->GetFourcc() == fourcc &&
-      request_context->camera_buffer->GetWidth() ==
-          stream_buffer->stream->width &&
-      request_context->camera_buffer->GetHeight() ==
-          stream_buffer->stream->height) {
-    // If no format conversion needs to be applied, directly copy the data over.
-    memcpy(output_frame.GetData(), request_context->camera_buffer->GetData(),
-           request_context->camera_buffer->GetDataSize());
-  } else {
-    // Perform the format conversion.
-    arc::CachedFrame cached_frame;
-    cached_frame.SetSource(request_context->camera_buffer.get(), 0);
-    cached_frame.Convert(request_context->request->settings, &output_frame);
-  }
-
-  request_context->request.reset();
-  // Mark the buffer as not in flight.
-  request_context->active = false;
-  return 0;
-}
-
-int V4L2Wrapper::GetInFlightBufferCount() {
-  int count = 0;
-  std::lock_guard<std::mutex> guard(buffer_queue_lock_);
-  for (auto& buffer : buffers_) {
-    if (buffer.active) {
-      count++;
-    }
-  }
-  return count;
 }
 
 }  // namespace v4l2_camera_hal
